+++ CodeStarterCamp_Week3/CoffeeShop.swift
@@ -12,15 +12,14 @@ final class CoffeeShop{
     private(set) var address: String
     private(set) var phoneNumber: String
     private var sales: Int
-    private var pickUpTable: [String: Coffee]
+    private var pickUpTable: [String: Coffee] = [:]
     private(set) var menus: [Coffee: Int]
     
-    init(name: String, address: String, phoneNumber: String, sales: Int, pickUpTable: [String : Coffee], menus: [Coffee : Int]) {
+    init(name: String, address: String, phoneNumber: String, sales: Int, menus: [Coffee : Int]) {
         self.name = name
         self.address = address
         self.phoneNumber = phoneNumber
         self.sales = sales
-        self.pickUpTable = pickUpTable
         self.menus = menus
     }
     
@@ -31,5 +30,4 @@ final class CoffeeShop{
         pickUpTable[name] = coffee
         print("\(name) 님이 주문하신 \(coffee.name)(이/가) 준비되었습니다. 픽업대에서 가져가주세요.")
     }
-
 }
